+++ src/pyiem/database.py
@@ -1,0 +1,106 @@
+"""Database helpers."""
+# stdlib
+from contextlib import contextmanager
+import getpass
+import warnings
+
+# third party
+import psycopg2
+from sqlalchemy import create_engine
+
+# NB: Careful of cyclic imports here...
+
+# Map system users back to something supported by akrherz/iem-database repo
+USERNAME_MAPPER = {
+    "apache": "nobody",
+    "www-data": "nobody",
+    "akrherz": "mesonet",
+    "meteor_ldm": "ldm",
+}
+
+
+def get_dbconnstr(name, **kwargs) -> str:
+    """Create a database connection string/URI.
+
+    Args:
+      name (str): the database name to connect to.
+      **kwargs: any additional arguments to pass to psycopg2.connect
+        user (str): the database user to connect as
+        host (str): the database host to connect to
+        port (int): the database port to connect to
+        connect_timeout (int): Connection timeout in seconds, default 30.
+    Returns:
+      str
+    """
+    user = kwargs.get("user")
+    if user is None:
+        user = USERNAME_MAPPER.get(getpass.getuser(), getpass.getuser())
+    host = kwargs.get("host")
+    if host is None:
+        host = f"iemdb-{name}.local"
+    port = kwargs.get("port")
+    if port is None:
+        port = 5432
+
+    # 15 seconds found to be a bit tight for local ISU congestion
+    return (
+        f"postgresql://{user}@{host}:{port}/{name}?"
+        f"connect_timeout={kwargs.get('connect_timeout', 30)}&"
+        f"gssencmode={kwargs.get('gssencmode', 'disable')}&"
+    )
+
+
+def get_dbconn(database="mesosite", user=None, host=None, port=5432, **kwargs):
+    """Helper function with business logic to get a database connection
+
+    Note that this helper could return a read-only database connection if the
+    connection to the primary server fails.
+
+    Args:
+      database (str,optional): the database name to connect to.
+        default: mesosite
+      user (str,optional): hard coded user to connect as, default: current user
+      host (str,optional): hard coded hostname to connect as,
+        default: iemdb.local
+      port (int,optional): the TCP port that PostgreSQL is listening
+        defaults to 5432
+      password (str,optional): the password to use.
+
+    Returns:
+      psycopg2 database connection
+    """
+    dsn = get_dbconnstr(database, user=user, host=host, port=port, **kwargs)
+    attempt = 0
+    conn = None
+    while attempt < 3:
+        attempt += 1
+        try:
+            conn = psycopg2.connect(dsn)
+        except (psycopg2.ProgrammingError, psycopg2.OperationalError) as exp:
+            if attempt == 3:
+                raise exp
+            warnings.warn(
+                f"database connection failure: {exp}, trying again",
+                stacklevel=2,
+            )
+    return conn
+
+
+@contextmanager
+def get_sqlalchemy_conn(text, **kwargs):
+    """An auto-disposing sqlalchemy context-manager helper.
+
+    This is used for when we really do not want to manage having pools of
+    database connections open.  So this isn't something that is fast!
+
+    Args:
+        text (str): the database to connect to, passed to get_dbconnstr
+        **kwargs: any additional arguments to pass to get_dbconnstr
+    """
+    engine = create_engine(get_dbconnstr(text, **kwargs))
+    try:
+        # Unsure if this is trouble or not.
+        with engine.connect() as conn:
+            yield conn
+    finally:
+        engine.dispose()