+++ src/pyiem/tracker.py
@@ -6,9 +6,10 @@ from email.mime.text import MIMEText
 try:
     from zoneinfo import ZoneInfo
 except ImportError:
-    from backports.zoneinfo import ZoneInfo
+    from backports.zoneinfo import ZoneInfo  # type:ignore
 
-from pyiem.util import get_dbconn
+from pyiem.database import get_dbconn
+from pyiem.util import LOG
 
 
 class TrackerEngine:
@@ -33,21 +34,24 @@ class TrackerEngine:
         self.action_count = 0
         self.emails = {}
 
-    def send_emails(self, really_send=True):
+    def send_emails(self):
         """Send out those SPAM emails!"""
         # Don't do anything if we have exceeded maxoffline
         if self.action_count >= self.maxactions and self.maxactions > 0:
             return
-        if not really_send:
-            return
-        s = smtplib.SMTP()
-        s.connect()
-        for email in self.emails:
-            msg = MIMEText(self.emails[email]["body"])
-            msg["From"] = "akrherz@iastate.edu"
-            msg["Subject"] = self.emails[email]["subject"]
-            s.sendmail(msg["From"], email, msg.as_string())
-        s.close()
+        with smtplib.SMTP() as s:
+            try:
+                s.connect()
+            except Exception as exp:
+                LOG.warning("smtp connection failed with %s", exp)
+            for email, entry in self.emails.items():
+                msg = MIMEText(entry["body"])
+                msg["From"] = "akrherz@iastate.edu"
+                msg["Subject"] = entry["subject"]
+                try:
+                    s.sendmail(msg["From"], email, msg.as_string())
+                except Exception as exp:
+                    LOG.warning("smtp sendmail failed with %s", exp)
 
     def offline_logic(self, sid, ob, pnetwork, nt):
         """offline logic
@@ -67,11 +71,8 @@ class TrackerEngine:
             (pnetwork, sid),
         )
         for row in self.pcursor:
-            open_tickets += (" %-6s %16s     %s\n" "") % (
-                row[0],
-                row[1].strftime("%Y-%m-%d %I %p"),
-                row[2],
-            )
+            dstr = row[1].strftime("%Y-%m-%d %I %p")
+            open_tickets += f" {row[0]:6.0f} {dstr:16s}     {row[2]}\n"
         # Get a listing of past 4 closed tickets
         closed_tickets = ""
         self.pcursor.execute(
@@ -80,11 +81,8 @@ class TrackerEngine:
             (pnetwork, sid),
         )
         for row in self.pcursor:
-            closed_tickets += (" %-6s %16s     %s\n" "") % (
-                row[0],
-                row[1].strftime("%Y-%m-%d %I %p"),
-                row[2],
-            )
+            dstr = row[1].strftime("%Y-%m-%d %I %p")
+            closed_tickets += f" {row[0]:6.0f} {dstr:16s}     {row[2]}\n"
         if closed_tickets == "":
             closed_tickets = " --None-- "
         if open_tickets == "":
@@ -98,7 +96,7 @@ class TrackerEngine:
         trackerid = self.pcursor.fetchone()[0]
         # Create a tt_log entry
         lts = ob["valid"].astimezone(ZoneInfo(nt.sts[sid]["tzname"]))
-        msg = "Site Offline since %s" % (lts.strftime("%d %b %Y %H:%M %Z"),)
+        msg = f"Site Offline since {lts:%d %b %Y %H:%M %Z}"
         self.pcursor.execute(
             "INSERT into tt_log (portfolio, s_mid, author, status_c, "
             "comments, tt_id) VALUES (%s, %s, %s, %s, %s, %s)",
@@ -160,9 +158,7 @@ class TrackerEngine:
         """
         trackerid = offline[sid]["trackerid"]
         # Create Log Entry
-        cmt = ("Site Back Online at: %s" "") % (
-            ob["valid"].strftime("%Y-%m-%d %H:%M:%S"),
-        )
+        cmt = f"Site Back Online at: {ob['valid']:%Y-%m-%d %H:%M:%S}"
         self.pcursor.execute(
             "INSERT into tt_log (portfolio, s_mid, author, status_c, "
             "comments, tt_id) VALUES (%s, %s, %s, %s, %s, %s)",
@@ -184,7 +180,7 @@ class TrackerEngine:
         days = delta.days
         hours = delta.seconds / 3600.0
         minutes = (delta.seconds % 3600) / 60.0
-        duration = "%.0f days %.0f hours %.0f minutes" % (days, hours, minutes)
+        duration = f"{days:.0f} days {hours:.0f} hours {minutes:.0f} minutes"
         mailstr = f"""
    ---------------------------------
    |  *** IEM TRACKER REPORT ***   |
@@ -215,7 +211,7 @@ IEM Tracker Action:  This trouble ticket has been marked
         for row in self.pcursor:
             email = row[0].lower()
             if email not in self.emails:
-                subject = ("[IEM] %s Online" "") % (nt.sts[sid]["name"],)
+                subject = f"[IEM] {nt.sts[sid]['name']} Online"
                 self.emails[email] = {"subject": subject, "body": mailstr}
             else:
                 subject = "[IEM] Multiple Sites"